bigint to ulong

+++ Crypto/Domain/Entities/Address.cs
@@ -2,7 +2,6 @@
 using Binebase.Exchange.CryptoService.Domain.Enums;
 using System;
 using System.Collections.Generic;
-using System.Numerics;
 
 namespace Binebase.Exchange.CryptoService.Domain.Entities
 {
@@ -12,10 +11,10 @@ namespace Binebase.Exchange.CryptoService.Domain.Entities
         public Currency Currency { get; set; }
         public AddressType Type { get; set; }
         public string Public { get; set; }
-        public BigInteger Balance { get; set; }
-        public int Index { get; set; }
-        public long GeneratedBlock { get; set; }
-        public long LastBlock { get; set; }
+        public ulong Balance { get; set; }
+        public uint Index { get; set; }
+        public ulong GeneratedBlock { get; set; }
+        public ulong LastBlock { get; set; }
         public virtual ICollection<Transaction> Transactions { get; } = new HashSet<Transaction>();
     }
 }
